+++ SFDS/src/main/java/dev/meres/sfds/Spaceship.kt
@@ -1,0 +1,6 @@
+package dev.meres.sfds
+
+data class Spaceship(
+    val name: String,
+    val communicationSystem: SFDS<String,String,String,String>
+)
