--- SFDS/src/main/java/dev/meres/sfds/const/ConnectionType.kt
@@ -1,5 +1,0 @@
-package dev.meres.sfds.const
-
-enum class ConnectionType {
-    BROADCAST, MULTICAST, UNICAST
-}
