--- SFDS/src/main/java/dev/meres/sfds/const/TransmissionBehavior.kt
@@ -1,5 +1,0 @@
-package dev.meres.sfds.const
-
-enum class TransmissionBehavior {
-    SHARED, FULL_DUPLEX, CLIENT_SERVER
-}
